tela de login e cadastro

+++ javascript/Aula 7 - Revisão-Projeto/Projeto Tela de Login/projeto-pizzaria/javascript/cadastro.js
@@ -1,12 +1,17 @@
+let textContent = document.getElementById('text-content')
 const cad = () => {
     const name = document.querySelector('input#nome').value;
     const sname = document.querySelector('input#sobrenome').value;
     const user = document.querySelector('input#usuario').value;
     const password = document.querySelector('input#senha').value;
+    const confirmPassword = document.querySelector('input#senha2').value
 
     try {
         if(!name || !sname || !user || !password){
             throw new Error('Todos os campos precisam ser preenchidos!')
+        }
+        if(password !== confirmPassword){
+            throw new Error('Senhas não conferem!')
         }
         
         localStorage.setItem("name", name);
@@ -16,7 +21,7 @@ const cad = () => {
         
         location.href = "../projeto-pizzaria/login.html";
     } catch (error) {
-        alert(error.message)
+        textContent.innerText = (error.message)
     }
 };
